test(router): consolidate issue 8115 coverage
